requirements.txt
torch==2.2.2
torchvision==0.17.2
torchsummary==1.5.1
numpy==1.26.4
pandas==2.2.2
opencv-python==4.11.0.86
pillow==10.2.0
matplotlib==3.8.0
plotly==5.20.0
imbalanced-learn==0.13.0
scikit-learn==1.3.2

